moved config.js to the application root and adjusted dependency files accordingly
